ch(session) add session request endpoint
add api endpoint for request a session

[Finishes #168325702]
