+++ Server/v1/routes/sesionRoute.js
@@ -4,11 +4,9 @@ import userAuth from '../middleware/authUser';
 import validateSession from '../helpers/sessionValidation';
 
 const router = express.Router();
-
+// session validation
 const { validateSessionRequest } = validateSession;
 
 router.post('/', validateSessionRequest, userAuth, sessionController.createSession);
-router.patch('/:sessionId/accept', userAuth, sessionController.acceptSession);
-router.patch('/:sessionId/reject', userAuth, sessionController.rejectSession);
 
 export default router;
